+++ weblate/wladmin/forms.py
@@ -8,6 +8,7 @@ from django.forms.widgets import MultiWidget
 from django.utils.translation import gettext_lazy
 
 from weblate.accounts.forms import EmailField
+from weblate.utils.validators import DomainOrIPValidator
 from weblate.wladmin.models import BackupService
 
 
@@ -24,7 +25,10 @@ class ActivateForm(forms.Form):
 
 class SSHAddForm(forms.Form):
     host = forms.CharField(
-        label=gettext_lazy("Hostname"), required=True, max_length=400
+        label=gettext_lazy("Hostname"),
+        required=True,
+        max_length=400,
+        validators=[DomainOrIPValidator()],
     )
     port = forms.IntegerField(
         label=gettext_lazy("Port"), required=False, min_value=1, max_value=65535
@@ -34,8 +38,6 @@ class SSHAddForm(forms.Form):
         super().__init__(*args, **kwargs)
         self.helper = FormHelper(self)
         self.helper.form_tag = False
-        self.helper.form_class = "form-inline"
-        self.helper.field_template = "bootstrap5/layout/inline_field.html"
 
 
 class TestMailForm(forms.Form):